add useSongLyrics
