+++ src/components/organismus/SongLyricsContainer.tsx
@@ -1,30 +1,15 @@
-import { FC, useEffect, useState } from 'react';
+import { FC } from 'react';
 import { Flex } from '@chakra-ui/react';
-import { getSong, getFurigana } from 'service/api';
+import { useSongLyrics } from 'hooks/useSongLyrics';
 import { SongHeader } from 'components/atoms/SongHeader';
 import { AlternatelyLyrics } from 'components/molecules/AlternatelyLyrics';
-import { Morpheme } from 'types/Morphology';
-import Song from 'types/Song';
 
 type Props = {
   songId: string;
 };
 
 const SongLyricsContainer: FC<Props> = ({ songId }) => {
-  const [song, setSong] = useState<Song | null>(null);
-  const [lyrics, setLyrics] = useState<Morpheme[]>([]);
-  if (songId == null) throw Error();
-  useEffect(() => {
-    const get = async () => {
-      const controller = new AbortController();
-      const { signal } = controller;
-      const { song } = await getSong({ id: songId }, { signal });
-      setSong(song);
-      const { morphemes } = await getFurigana({ text: song.lyrics });
-      setLyrics(morphemes);
-    };
-    get();
-  }, [songId]);
+  const [{ song, lyrics }] = useSongLyrics(songId);
 
   if (song == null) return <></>;
